feat: Implemented route to delete posts/files

+++ models/post.js
@@ -58,20 +58,38 @@ const delistPost = async (post) => {
     await User.updateOne({ _id: userId }, { $pull: { posts: post._id } })
 }
 
-// Deletes a post and its associated file
-const deletePost = async (postId) => {
-    const post = await Post.findOne({ _id: postId })
+// Checks if the given user has permission to edit/delete a given post
+const hasPermission = (post, user) => {
+    if (user) {
+        console.log(user._id.equals(post.createdBy))
+        if (user._id.equals(post.createdBy) || user.role === 'Admin') {
+            return true
+        }
+        return false
+    }
+    return false
+}
+
+// Deletes a post and its associated file, checking if the associated user has permission to delete that post
+const deletePost = async (postId, user) => {
     try {
-        const fileId = post.fileId
-        if (fileId) {
-            await db.deleteFile(fileId)
-            await Post.deleteOne({ _id: postId })
-            await delistPost(post)
-        } else {
-            console.log('Error: Failed to find post')
+        const post = await Post.findOne({ _id: postId })
+        if (hasPermission(post, user)) {
+            try {
+                const fileId = post.fileId
+                if (fileId) {
+                    await db.deleteFile(fileId)
+                    await Post.deleteOne({ _id: postId })
+                    await delistPost(post)
+                } else {
+                    console.log('Error: Failed to find post')
+                }
+            } catch (err) {
+                console.log('Error: Failed to delete post')
+            }
         }
     } catch (err) {
-        console.log('Error: Failed to delete post')
+        console.log(err)
     }
 }
